feat: наработки api удаления счету(2)

+++ front/src/main/java/org/ivanov/front/service/impl/WalletServiceImpl.java
@@ -5,7 +5,12 @@ import org.ivanov.accountdto.wallet.CreateWalletDto;
 import org.ivanov.accountdto.wallet.ResponseWalletDto;
 import org.ivanov.front.client.AccountClient;
 import org.ivanov.front.service.WalletService;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
+
+import java.math.BigDecimal;
+import java.util.Set;
 
 @Service
 @RequiredArgsConstructor
@@ -13,12 +18,20 @@ public class WalletServiceImpl implements WalletService {
     private final AccountClient client;
 
     @Override
-    public ResponseWalletDto createWallet(Long accountId, CreateWalletDto createWalletDto) {
+    public ResponseWalletDto createWallet(Long accountId, CreateWalletDto createWalletDto, Set<ResponseWalletDto> wallets) {
+        boolean isValidWallet = wallets.stream().noneMatch(w -> w.walletType().equals(createWalletDto.walletType()));
+        if (!isValidWallet) {
+            throw new ResponseStatusException(HttpStatus.CONFLICT, "Счет этого типа уже существует");
+        }
         return client.createWallet(accountId, createWalletDto);
     }
 
     @Override
-    public void deleteWallet(Long accountId, Long walletId) {
+    public void deleteWallet(Long accountId, Long walletId, Set<ResponseWalletDto> wallets) {
+        boolean isValidWallet = wallets.stream().noneMatch(w -> w.balance().compareTo(BigDecimal.ZERO) == 0);
+        if (!isValidWallet) {
+            throw new ResponseStatusException(HttpStatus.CONFLICT, "Нельзя удалить ненулевой счет");
+        }
         client.deleteWallet(accountId, walletId);
     }
 }
